+++ tests/functional_tests/test_general_formatting.py
@@ -96,6 +96,7 @@ from tests.util import check_formatting, read_test_data
         "unformatted/411_create_clone.sql",
         "unformatted/412_pragma.sql",
         "unformatted/900_create_view.sql",
+        "unformatted/998_unsupported_ddl_with_jinja.sql",
         "unformatted/999_unsupported_ddl.sql",
     ],
 )